+++ Assets/VRM10/Runtime/Components/Vrm10Runtime/Vrm10RuntimeLookAt.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 08fd3fec20846764dbb598cf75b13356
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
